pyproject.toml
[tool.poetry]
name = "vnguide_v2"
version = "0.1.0"
description = ""
authors = ["dungpq <phanquangdng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
beanie = "^1.13.0"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
python-decouple = "^3.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
inflect = "^6.0.2"
schedule = "^1.1.0"
aiohttp = "^3.8.3"
aiodns = "^3.0.0"
xmltodict = "^0.13.0"
html-to-json = "^2.0.0"
python-dateutil = "^2.8.2"
pytz = "^2022.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
